pyproject.toml
[tool.poetry]
name = "py-cw-5-db"
version = "0.1.0"
description = ""
authors = ["KonstantinMikheev <mikhey17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
